requirements.txt
feedparser==6.0.11
requests==2.31.0
PyYAML==6.0.1
beautifulsoup4==4.12.2
python-dateutil==2.8.2
